requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.2
Pillow>=8.0.0
tifffile>=2021.7.2
PyYAML>=5.3.1
tqdm>=4.50.2
joblib>=0.17.0
python-box>=6.0.2
matplotlib>=3.3.2
splitfolders>=0.4.3
scikit-image>=0.18.1
streamlit>=1.12.0